chatboot_ivana-main/requirements.txt
flask
flask-cors
langchain
langchain-groq
python-dotenv
langchain-huggingface
sentence-transformers
pypdf
faiss-cpu
fpdf2
